+++ client/src/App.tsx
@@ -4,6 +4,7 @@ import NotFound from "@/pages/NotFound";
 import { Route, Switch } from "wouter";
 import ErrorBoundary from "./components/ErrorBoundary";
 import { ThemeProvider } from "./contexts/ThemeContext";
+import { ResultsProvider } from "./contexts/ResultsContext";
 import JobDashboardEnhanced from "./pages/JobDashboardEnhanced";
 import IntelligentNormalization from "./pages/IntelligentNormalization";
 import TestCredentials from "./pages/TestCredentials";
@@ -36,10 +37,12 @@ function App() {
         defaultTheme="light"
         switchable
       >
-        <TooltipProvider>
-          <Toaster />
-          <Router />
-        </TooltipProvider>
+        <ResultsProvider>
+          <TooltipProvider>
+            <Toaster />
+            <Router />
+          </TooltipProvider>
+        </ResultsProvider>
       </ThemeProvider>
     </ErrorBoundary>
   );